backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9    
paho-mqtt>=1.6.1       
python-dotenv>=1.0.0
networkx